+++ py_bootstrap/files_processors/generate.py
@@ -2,6 +2,7 @@ __all__ = ("GenerateFilesProcessor",)
 
 import logging
 import typing as t
+from functools import cached_property
 from pathlib import Path
 
 from .copy import CopyFilesProcessor
@@ -15,41 +16,28 @@ logger = logging.getLogger(__name__)
 
 class GenerateFilesProcessor(CopyFilesProcessor):
     _context: dict[str, str]
-    _entry_point_file_name: str
+
+    @cached_property
+    def _descriptor_file_name(self) -> str:
+        return f"{self._source_path}.py"
 
     def set_context(self, value: dict[str, str]):
         self._context = value
 
-    def set_entry_point_file_name(self, value: str):
-        self._entry_point_file_name = value
-
     def generate_content_from_template(self, template: str) -> str:
         return template.format(**self._context)
 
-    def process_directory(self, rel_path: "Path", dir_name: str):
-        path = self._destination_path.joinpath(rel_path, dir_name)
-        path = Path(
-            self.generate_content_from_template(template=path.as_posix())
-        )
-        path.mkdir(parents=True, exist_ok=True)
-
-    def check_file_for_processing(
-        self, rel_path: "Path", file_name: str
-    ) -> bool:
-        if file_name == self._entry_point_file_name:
+    def check_file_for_processing(self, rel_path: "Path", file_name: str) -> bool:
+        if file_name == self._descriptor_file_name:
             return False
-        return super().check_file_for_processing(
-            rel_path=rel_path, file_name=file_name
-        )
+        return super().check_file_for_processing(rel_path=rel_path, file_name=file_name)
 
     def process_file(self, rel_path: "Path", file_name: str):
         if file_name.endswith(".tmpl"):
             source_path = self._source_path.joinpath(rel_path, file_name)
 
             file_name = file_name.removesuffix(".tmpl")
-            destination_path = self._destination_path.joinpath(
-                rel_path, file_name
-            )
+            destination_path = self._destination_path.joinpath(rel_path, file_name)
             destination_path = Path(
                 self.generate_content_from_template(
                     template=destination_path.as_posix()
@@ -60,6 +48,18 @@ class GenerateFilesProcessor(CopyFilesProcessor):
             prepared_content = self.generate_content_from_template(
                 template=origin_content
             )
-            destination_path.write_text(prepared_content)
+            if self._dry_run:
+                self.log_dry_run_action(
+                    action="generate file %r from %r template"
+                    % (destination_path, source_path)
+                )
+            else:
+                destination_path.write_text(prepared_content)
+                logger.info(
+                    "%r. generate file %r from %r template",
+                    self,
+                    destination_path,
+                    source_path,
+                )
         else:
             super().process_file(rel_path=rel_path, file_name=file_name)
